move fixtures to lib
