+++ packages/components/lib/Label/Label.fixture.tsx
@@ -1,6 +1,6 @@
 import { useValue } from "react-cosmos/client";
 
-import { Label } from "../../lib/main";
+import { Label } from "../main";
 
 export default function LabelFixture() {
   const [content] = useValue("Content", { defaultValue: "This is a label" });
