+++ packages/components/lib/fixtures/simpleForm.fixture.tsx
@@ -1,6 +1,6 @@
 import { useValue } from "react-cosmos/client";
 
-import { Button, FormField, Stack, TextInput } from "../../../lib/main";
+import { Button, FormField, Stack, TextInput } from "../main";
 
 export default function ButtonFixture() {
   const [label] = useValue("Label", { defaultValue: "Blick me!" });
